Create a module for whitelisted target features.
