+++ src/librustc_codegen_llvm/attributes.rs
@@ -17,6 +17,7 @@ use rustc::session::Session;
 use rustc::session::config::Sanitizer;
 use rustc::ty::TyCtxt;
 use rustc::ty::query::Providers;
+use rustc_codegen_utils::target_features::all_known_features;
 use rustc_data_structures::sync::Lrc;
 use rustc_data_structures::fx::FxHashMap;
 use rustc_target::spec::PanicStrategy;
@@ -268,7 +269,7 @@ pub fn provide(providers: &mut Providers) {
         if tcx.sess.opts.actually_rustdoc {
             // rustdoc needs to be able to document functions that use all the features, so
             // whitelist them all
-            Lrc::new(llvm_util::all_known_features()
+            Lrc::new(all_known_features()
                 .map(|(a, b)| (a.to_string(), b.map(|s| s.to_string())))
                 .collect())
         } else {
